Fixed bugs in operator and number buttons
Pressing an operator button twice would take the first number
entered as both first and second and calculate accordingly. Set it so
hitting an operator twice in a row does nothing. Also added boolean
variable to check if we have pressed the equals button so we know to
overwrite the display if so upon pressing number button.