Add keyboard support
Pressing a key on the keyboard that corresponds with a button on the
calculator will click that button. As we use enter for =, also have set
it so that buttons can't be tabbed too and used preventDefault on
mousedown events to prevent focusing on a button after clicking and
thus it being clicked by pressing enter.